+++ packages/lib/src/calculators/calculate-tip-result.spec.ts
@@ -1,10 +1,7 @@
 import { describe, expect, test } from 'vitest';
-import { Tip } from '../tip';
-import {
-  calculateTipResult,
-  tipRuleDescriptions,
-  TipRule,
-} from './calculate-tip-result';
+import { tipRuleDescriptions, TipRuleId } from '../model/rules/tip-rule';
+import { Tip } from '../model/tip';
+import { calculateTipResult } from './calculate-tip-result';
 
 // Passing in an unreal points value in order to test object identity
 function makeTipMock(
@@ -84,7 +81,7 @@ const mockData = [
 ];
 
 describe('Tippberechnung nach Regeln: ' + tipRuleDescriptions[0].name, () => {
-  const rule: TipRule = 'DreiOderEinPunktJokerVerdoppelt';
+  const ruleId: TipRuleId = 'drei-oder-ein-punkt-joker-verdoppelt';
 
   const expectedPoints = [3, 6, 1, 2, 1, 2, 0, 0, 0, 0, 0, 0];
 
@@ -92,7 +89,7 @@ describe('Tippberechnung nach Regeln: ' + tipRuleDescriptions[0].name, () => {
     mockData.forEach((match) => {
       match.tips.forEach((tip, ix) => {
         const t = makeTipMock(tip.tip, tip.joker);
-        const calculatedTip = calculateTipResult(t, match.result, rule);
+        const calculatedTip = calculateTipResult(t, match.result, ruleId);
         expect(calculatedTip).not.toBe(t);
         expect(calculatedTip.points).toBe(expectedPoints[ix]);
       });
@@ -101,34 +98,34 @@ describe('Tippberechnung nach Regeln: ' + tipRuleDescriptions[0].name, () => {
 
   test('ändert bei einer Neuberechnung nichts', () => {
     const tip = makeTipMock('2:1', true, 6);
-    const calculatedTip = calculateTipResult(tip, '2:1', rule);
+    const calculatedTip = calculateTipResult(tip, '2:1', ruleId);
     expect(calculatedTip).toBe(tip);
   });
 
   test('setzt alle extra Flags zurück', () => {
     const tip = makeTipMock('2:1', true, 6, true);
-    const calculatedTip = calculateTipResult(tip, '2:1', rule);
+    const calculatedTip = calculateTipResult(tip, '2:1', ruleId);
     expect(calculatedTip.lonelyHit).toBeUndefined();
     expect(calculatedTip).not.toBe(tip);
   });
 
   test('setzt Joker auf leerem Tipp zurück', () => {
     const tip = makeTipMock('', true, 0);
-    const calculatedTip = calculateTipResult(tip, '2:1', rule);
+    const calculatedTip = calculateTipResult(tip, '2:1', ruleId);
     expect(calculatedTip.joker).toBeFalsy();
     expect(calculatedTip).not.toBe(tip);
   });
 
   test('nullt die Punkte bei leerem Match', () => {
     const tip = makeTipMock('2:1', false, 3);
-    const calculatedTip = calculateTipResult(tip, '', rule);
+    const calculatedTip = calculateTipResult(tip, '', ruleId);
     expect(calculatedTip.points).toBe(0);
     expect(calculatedTip).not.toBe(tip);
   });
 });
 
 describe('Tippberechnung nach Regeln: ' + tipRuleDescriptions[1].name, () => {
-  const rule: TipRule = 'DreiZweiOderEinPunktJokerVerdoppelt';
+  const ruleId: TipRuleId = 'drei-zwei-oder-ein-punkt-joker-verdoppelt';
 
   const expectedPoints = [
     [3, 6, 1, 2, 2, 4, 0, 0, 0, 0, 0, 0],
@@ -140,7 +137,7 @@ describe('Tippberechnung nach Regeln: ' + tipRuleDescriptions[1].name, () => {
     mockData.forEach((match, ixMatch) => {
       match.tips.forEach((tip, ixTip) => {
         const t = makeTipMock(tip.tip, tip.joker);
-        const calculatedTip = calculateTipResult(t, match.result, rule);
+        const calculatedTip = calculateTipResult(t, match.result, ruleId);
         expect(calculatedTip).not.toBe(t);
         expect(calculatedTip.points).toBe(expectedPoints[ixMatch][ixTip]);
       });
@@ -149,27 +146,27 @@ describe('Tippberechnung nach Regeln: ' + tipRuleDescriptions[1].name, () => {
 
   test('ändert bei einer Neuberechnung nichts', () => {
     const tip = makeTipMock('2:1', true, 6);
-    const calculatedTip = calculateTipResult(tip, '2:1', rule);
+    const calculatedTip = calculateTipResult(tip, '2:1', ruleId);
     expect(calculatedTip).toBe(tip);
   });
 
   test('setzt alle extra Flags zurück', () => {
     const tip = makeTipMock('2:1', true, 6, true);
-    const calculatedTip = calculateTipResult(tip, '2:1', rule);
+    const calculatedTip = calculateTipResult(tip, '2:1', ruleId);
     expect(calculatedTip.lonelyHit).toBeUndefined();
     expect(calculatedTip).not.toBe(tip);
   });
 
   test('setzt Joker auf leerem Tipp zurück', () => {
     const tip = makeTipMock('', true, 0);
-    const calculatedTip = calculateTipResult(tip, '2:1', rule);
+    const calculatedTip = calculateTipResult(tip, '2:1', ruleId);
     expect(calculatedTip.joker).toBeFalsy();
     expect(calculatedTip).not.toBe(tip);
   });
 
   test('nullt die Punkte bei leerem Match', () => {
     const tip = makeTipMock('2:1', false, 3);
-    const calculatedTip = calculateTipResult(tip, '', rule);
+    const calculatedTip = calculateTipResult(tip, '', ruleId);
     expect(calculatedTip.points).toBe(0);
     expect(calculatedTip).not.toBe(tip);
   });