+++ packages/lib/src/index.ts
@@ -1,6 +1,23 @@
 // Model
+export * from './model/base/model';
+export * from './model/rules/rule';
+export * from './model/rules/tip-rule';
+export * from './model/rules/match-rule';
+export * from './model/rules/round-rule';
+export * from './model/rules/extra-question-rule';
 export * from './model/championship';
+export * from './model/championchip-rules';
+export * from './model/championship';
+export * from './model/tip';
+export * from './model/match';
+
+// Calculators
+export * from './calculators/calculate-tip-result';
+export * from './calculators/calculate-match-results';
 
 // Repository
 export * from './repository/collection';
 export * from './repository/constraint';
+export * from './repository/create-entity';
+export * from './repository/patch-entity';
+export * from './repository/update-entity';