+++ packages/lib/src/model/rules/round-rule.ts
@@ -1,0 +1,17 @@
+import { Rule } from './rule';
+
+export type RoundRuleId = 'keine-besonderheiten';
+
+export interface RoundRule extends Rule {
+  id: RoundRuleId;
+}
+
+export const roundRuleDescriptions: RoundRule[] = [
+  {
+    id: 'keine-besonderheiten',
+    name: 'Keine Besonderheiten',
+    description: `
+    Es gibt keine Sonderregeln zum Abschluss einer Runde.
+  `,
+  },
+];